refactor(travel): isolate command transformation
